+++ docs/News.rst
@@ -13,6 +13,8 @@ SQLObject 3.4.0 (master)
 * Python 2.6 is no longer supported. The minimal supported version is
   Python 2.7.
 
+* Encode/decode binary values for PyGreSQL driver.
+
 SQLObject 3.3.0
 ===============